refactor: export createT function
