+++ examples/next/locales/index.ts
@@ -1,7 +1,9 @@
 import { createI18n } from 'next-international';
 import type Locale from './en';
 
-export const { useI18n, I18nProvider, useChangeLocale, defineLocale, getLocaleProps } = createI18n<typeof Locale>({
+export const { useI18n, useScopedI18n, I18nProvider, useChangeLocale, defineLocale, getLocaleProps } = createI18n<
+  typeof Locale
+>({
   en: () => import('./en'),
   fr: () => import('./fr'),
 });
