+++ packages/next-international/src/i18n/create-use-scoped-i18n.ts
@@ -1,14 +1,7 @@
-import React, { useContext, Context, isValidElement, cloneElement, ReactNode } from 'react';
-import type {
-  BaseLocale,
-  LocaleKeys,
-  LocaleValue,
-  Params,
-  ParamsObject,
-  ScopedValue,
-  Scopes,
-} from 'international-types';
-import type { LocaleContext, ReactParamsObject, LocaleMap } from '../types';
+import { useContext, Context } from 'react';
+import type { BaseLocale, Scopes } from 'international-types';
+import type { LocaleContext } from '../types';
+import { createT } from './create-t';
 
 export function createScopedUsei18n<Locale extends BaseLocale>(I18nContext: Context<LocaleContext<Locale> | null>) {
   return function useScopedI18n<Scope extends Scopes<Locale> | undefined>(scope: Scope) {
@@ -18,58 +11,6 @@ export function createScopedUsei18n<Locale extends BaseLocale>(I18nContext: Cont
       throw new Error('`useI18n` must be used inside `I18nProvider`');
     }
 
-    function createT() {
-      function t<Key extends LocaleKeys<Locale, Scope>, Value extends LocaleValue = ScopedValue<Locale, Scope, Key>>(
-        key: Key,
-        ...params: Params<Value>['length'] extends 0 ? [] : [ParamsObject<Value>]
-      ): string;
-      function t<Key extends LocaleKeys<Locale, Scope>, Value extends LocaleValue = ScopedValue<Locale, Scope, Key>>(
-        key: Key,
-        ...params: Params<Value>['length'] extends 0 ? [] : [ReactParamsObject<Value>]
-      ): React.ReactNode;
-      function t<Key extends LocaleKeys<Locale, Scope>, Value extends LocaleValue = ScopedValue<Locale, Scope, Key>>(
-        key: Key,
-        ...params: Params<Value>['length'] extends 0 ? [] : [ParamsObject<Value> | ReactParamsObject<Value>]
-      ) {
-        const { localeContent, fallbackLocale } = context as LocaleContext<Locale>;
-
-        const value = (
-          (scope ? localeContent[`${scope}.${key}`] : localeContent[key]) ||
-          (scope ? fallbackLocale?.[`${scope}.${key}`] : fallbackLocale?.[key]) ||
-          key
-        )?.toString();
-        const paramObject = params[0];
-
-        if (!paramObject) {
-          return value;
-        }
-
-        let isString = true;
-
-        const result = value?.split(/({[^}]*})/).map((part, index) => {
-          const match = part.match(/{(.*)}/);
-
-          if (match) {
-            const param = match[1] as keyof Locale;
-            const paramValue = (paramObject as LocaleMap<Locale>)[param];
-
-            if (typeof paramValue !== 'string') isString = false;
-
-            return isValidElement(paramValue)
-              ? cloneElement(paramValue, { key: `${String(param)}-${index}` })
-              : (paramValue as ReactNode);
-          }
-
-          // if there's no match - it's not a variable and just a normal string
-          return part;
-        });
-
-        return isString ? result?.join('') : result;
-      }
-
-      return t;
-    }
-
-    return createT();
+    return createT(context, scope);
   };
 }
